Changes for succesful compilation of most mex files. Changed to 64 bit ints where required, disabled non-fixable or missing methods.

+++ gp_toolbox_fast_marching/mex/perform_front_propagation_anisotropic.cpp
@@ -27,22 +27,22 @@ void mexFunction(int nlhs, mxArray *plhs[],
 	double *mask_data = mxGetPr(prhs[1]); 	// mask field
 	double alpha = mxGetScalar(prhs[2]); 	// alpha
 	double* start_points = mxGetPr(prhs[3]);	// starting points
-	int nstart = mxGetN(prhs[3]); // number of starting points
+	size_t nstart = mxGetN(prhs[3]); // number of starting points
 	double dmax = mxGetScalar(prhs[4]);	// starting points
 
 	if( mxGetM(prhs[3])!=3 )
 		mexErrMsgTxt("start_points should be of size (3,p).");
 	
     // width, height, depth
-    int w = mxGetDimensions(prhs[0])[0];
-	int h = mxGetDimensions(prhs[0])[1];
-	int d = mxGetDimensions(prhs[0])[2];
+    size_t w = mxGetDimensions(prhs[0])[0];
+	size_t h = mxGetDimensions(prhs[0])[1];
+	size_t d = mxGetDimensions(prhs[0])[2];
 
 	// number of dimensions for tensor field
-    const int ndims = 6;
+    const size_t ndims = 6;
 
 	// output
-	int dims[3] = {w, h, d};
+	size_t dims[3] = {w, h, d};
 	// distance map
     plhs[0] = mxCreateNumericArray(3, dims, mxDOUBLE_CLASS, mxREAL);
 	double *dist_data = mxGetPr(plhs[0]);
@@ -50,13 +50,13 @@ void mexFunction(int nlhs, mxArray *plhs[],
     plhs[1] = mxCreateNumericArray(3, dims, mxDOUBLE_CLASS, mxREAL);
     double *voro = mxGetPr(plhs[1]);
     
-    int size = w*h*d;
-    for(int i=0; i<size; i++) { dist_data[i] = DBL_MAX; voro[i] = 0; }
+    size_t size = w*h*d;
+    for(size_t i=0; i<size; i++) { dist_data[i] = DBL_MAX; voro[i] = 0; }
 		
 #if 0
 	double *tensor_power_data;
     tensor_power_data = reinterpret_cast<double*> (malloc(size*ndims*sizeof(double)));
-    for(int i=0; i<size*ndims; i++)
+    for(size_t i=0; i<size*ndims; i++)
 		tensor_power_data[i] = tensor_data[i];
 #endif
     
@@ -71,11 +71,11 @@ void mexFunction(int nlhs, mxArray *plhs[],
     int pos;
 	// set up starting points
     double labels = 1, nstartd = (double)nstart;
-	for( int i=0; i<nstart; ++i )
+	for( size_t i=0; i<nstart; ++i )
 	{
-		int x0 = start_points_(0,i);
-		int y0 = start_points_(1,i);
-		int z0 = start_points_(2,i);
+		size_t x0 = start_points_(0,i);
+		size_t y0 = start_points_(1,i);
+		size_t z0 = start_points_(2,i);
         pos = z0*w*h + y0*h + x0;
 		dist_data[pos] = 0;
         voro[pos] = labels; // /nstartd;
